split horizontal.py into geometry.py and horizontal.py
