+++ model/atmosphere/diffusion/tests/diffusion_tests/test_diffusion.py
@@ -11,7 +11,7 @@ import pytest
 
 from icon4py.model.atmosphere.diffusion import diffusion, diffusion_utils
 from icon4py.model.common.grid import vertical as v_grid
-from icon4py.model.common.grid.horizontal import CellParams, EdgeParams
+from icon4py.model.common.grid.geometry import CellParams, EdgeParams
 from icon4py.model.common.settings import backend
 from icon4py.model.common.test_utils import (
     datatest_utils as dt_utils,
