+++ model/atmosphere/dycore/src/icon4py/model/atmosphere/dycore/velocity/velocity_advection.py
@@ -10,6 +10,7 @@ import gt4py.next as gtx
 import numpy as np
 
 import icon4py.model.atmosphere.dycore.velocity.velocity_advection_program as velocity_prog
+import icon4py.model.common.grid.geometry
 from icon4py.model.atmosphere.dycore.add_extra_diffusion_for_normal_wind_tendency_approaching_cfl import (
     add_extra_diffusion_for_normal_wind_tendency_approaching_cfl,
 )
@@ -55,7 +56,7 @@ class VelocityAdvection:
         metric_state: solve_nh_states.MetricStateNonHydro,
         interpolation_state: solve_nh_states.InterpolationState,
         vertical_params: v_grid.VerticalGridParams,
-        edge_params: h_grid.EdgeParams,
+        edge_params: icon4py.model.common.grid.geometry.EdgeParams,
         owner_mask: fa.CellField[bool],
     ):
         self._initialized = False
