+++ model/atmosphere/dycore/tests/dycore_tests/test_solve_nonhydro.py
@@ -10,6 +10,7 @@ import logging
 
 import pytest
 
+import icon4py.model.common.grid.geometry
 from icon4py.model.atmosphere.dycore.nh_solve import solve_nonhydro as solve_nh
 from icon4py.model.atmosphere.dycore.state_utils import (
     states as solve_nh_states,
@@ -143,8 +144,12 @@ def test_nonhydro_predictor_step(
     interpolation_state = construct_interpolation_state_for_nonhydro(interpolation_savepoint)
     metric_state_nonhydro = construct_nh_metric_state(metrics_savepoint, icon_grid.num_levels)
 
-    cell_geometry: h_grid.CellParams = grid_savepoint.construct_cell_geometry()
-    edge_geometry: h_grid.EdgeParams = grid_savepoint.construct_edge_geometry()
+    cell_geometry: icon4py.model.common.grid.geometry.CellParams = (
+        grid_savepoint.construct_cell_geometry()
+    )
+    edge_geometry: icon4py.model.common.grid.geometry.EdgeParams = (
+        grid_savepoint.construct_edge_geometry()
+    )
 
     solve_nonhydro = solve_nh.SolveNonhydro()
     nlev = icon_grid.num_levels
@@ -569,8 +574,12 @@ def test_nonhydro_corrector_step(
     interpolation_state = construct_interpolation_state_for_nonhydro(interpolation_savepoint)
     metric_state_nonhydro = construct_nh_metric_state(metrics_savepoint, icon_grid.num_levels)
 
-    cell_geometry: h_grid.CellParams = grid_savepoint.construct_cell_geometry()
-    edge_geometry: h_grid.EdgeParams = grid_savepoint.construct_edge_geometry()
+    cell_geometry: icon4py.model.common.grid.geometry.CellParams = (
+        grid_savepoint.construct_cell_geometry()
+    )
+    edge_geometry: icon4py.model.common.grid.geometry.EdgeParams = (
+        grid_savepoint.construct_edge_geometry()
+    )
 
     solve_nonhydro = solve_nh.SolveNonhydro()
     solve_nonhydro.init(
@@ -757,8 +766,12 @@ def test_run_solve_nonhydro_single_step(
     interpolation_state = construct_interpolation_state_for_nonhydro(interpolation_savepoint)
     metric_state_nonhydro = construct_nh_metric_state(metrics_savepoint, icon_grid.num_levels)
 
-    cell_geometry: h_grid.CellParams = grid_savepoint.construct_cell_geometry()
-    edge_geometry: h_grid.EdgeParams = grid_savepoint.construct_edge_geometry()
+    cell_geometry: icon4py.model.common.grid.geometry.CellParams = (
+        grid_savepoint.construct_cell_geometry()
+    )
+    edge_geometry: icon4py.model.common.grid.geometry.EdgeParams = (
+        grid_savepoint.construct_edge_geometry()
+    )
 
     solve_nonhydro = solve_nh.SolveNonhydro()
     solve_nonhydro.init(
@@ -886,8 +899,12 @@ def test_run_solve_nonhydro_multi_step(
     interpolation_state = construct_interpolation_state_for_nonhydro(interpolation_savepoint)
     metric_state_nonhydro = construct_nh_metric_state(metrics_savepoint, icon_grid.num_levels)
 
-    cell_geometry: h_grid.CellParams = grid_savepoint.construct_cell_geometry()
-    edge_geometry: h_grid.EdgeParams = grid_savepoint.construct_edge_geometry()
+    cell_geometry: icon4py.model.common.grid.geometry.CellParams = (
+        grid_savepoint.construct_cell_geometry()
+    )
+    edge_geometry: icon4py.model.common.grid.geometry.EdgeParams = (
+        grid_savepoint.construct_edge_geometry()
+    )
 
     solve_nonhydro = solve_nh.SolveNonhydro()
     solve_nonhydro.init(
