+++ model/driver/src/icon4py/model/driver/test_cases/jablonowski_williamson.py
@@ -12,6 +12,7 @@ import pathlib
 
 import gt4py.next as gtx
 
+import icon4py.model.common.grid.geometry
 from icon4py.model.atmosphere.diffusion import diffusion_states as diffus_states
 from icon4py.model.atmosphere.dycore import init_exner_pr
 from icon4py.model.atmosphere.dycore.state_utils import states as solve_nh_states
@@ -36,8 +37,8 @@ log = logging.getLogger(__name__)
 
 def model_initialization_jabw(
     grid: icon_grid.IconGrid,
-    cell_param: h_grid.CellParams,
-    edge_param: h_grid.EdgeParams,
+    cell_param: icon4py.model.common.grid.geometry.CellParams,
+    edge_param: icon4py.model.common.grid.geometry.EdgeParams,
     path: pathlib.Path,
     rank=0,
 ) -> tuple[
