+++ requirements-dev-opt.txt
@@ -1,4 +1,4 @@
-git+https://github.com/ghex-org/GHEX.git@master#subdirectory=bindings/python
+git+https://github.com/ghex-org/GHEX.git@61ca5f0e3b6b473f9ca397052b2b0c1d8a4d3c55#subdirectory=bindings/python
 -r base-requirements-dev.txt
 # icon4py model
 -e ./model/common[all]
